requirements.txt
Flask~=3.1.0
Flask-Bcrypt~=1.0.1
Flask-JWT-Extended~=4.6.0
Flask-SQLAlchemy
Flask-Migrate
PyJWT~=2.10.0
SQLAlchemy~=2.0.36
werkzeug~=3.1.3
httpx~=0.28.1
requests~=2.32.3